trans: Move trans_foreign_mod and trans_impl to trans_item.
